requirements.txt
django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.2.2
setuptools>=65.0.0
setuptools